@@ -43,6 +43,10 @@ class Config {
   Declaration get structDecl => _structDecl;
   late Declaration _structDecl;
 
+  /// Declaration config for Unions.
+  Declaration get unionDecl => _unionDecl;
+  late Declaration _unionDecl;
+
   /// Declaration config for Enums.
   Declaration get enumClassDecl => _enumClassDecl;
   late Declaration _enumClassDecl;
@@ -77,8 +81,12 @@ class Config {
   late CommentType _commentType;
 
   /// Whether structs that are dependencies should be included.
-  StructDependencies get structDependencies => _structDependencies;
-  late StructDependencies _structDependencies;
+  CompoundDependencies get structDependencies => _structDependencies;
+  late CompoundDependencies _structDependencies;
+
+  /// Whether unions that are dependencies should be included.
+  CompoundDependencies get unionDependencies => _unionDependencies;
+  late CompoundDependencies _unionDependencies;
 
   /// Holds config for how struct packing should be overriden.
   StructPackingOverride get structPackingOverride => _structPackingOverride;
@@ -234,6 +242,15 @@ class Config {
           _structDecl = result as Declaration;
         },
       ),
+      [strings.unions]: Specification<Declaration>(
+        requirement: Requirement.no,
+        validator: declarationConfigValidator,
+        extractor: declarationConfigExtractor,
+        defaultValue: () => Declaration(),
+        extractedResult: (dynamic result) {
+          _unionDecl = result as Declaration;
+        },
+      ),
       [strings.enums]: Specification<Declaration>(
         requirement: Requirement.no,
         validator: declarationConfigValidator,
@@ -312,14 +329,23 @@ class Config {
         extractedResult: (dynamic result) =>
             _commentType = result as CommentType,
       ),
-      [strings.structs, strings.structDependencies]:
-          Specification<StructDependencies>(
+      [strings.structs, strings.dependencyOnly]:
+          Specification<CompoundDependencies>(
         requirement: Requirement.no,
-        validator: structDependenciesValidator,
-        extractor: structDependenciesExtractor,
-        defaultValue: () => StructDependencies.full,
+        validator: dependencyOnlyValidator,
+        extractor: dependencyOnlyExtractor,
+        defaultValue: () => CompoundDependencies.full,
         extractedResult: (dynamic result) =>
-            _structDependencies = result as StructDependencies,
+            _structDependencies = result as CompoundDependencies,
+      ),
+      [strings.unions, strings.dependencyOnly]:
+          Specification<CompoundDependencies>(
+        requirement: Requirement.no,
+        validator: dependencyOnlyValidator,
+        extractor: dependencyOnlyExtractor,
+        defaultValue: () => CompoundDependencies.full,
+        extractedResult: (dynamic result) =>
+            _unionDependencies = result as CompoundDependencies,
       ),
       [strings.structs, strings.structPack]:
           Specification<StructPackingOverride>(
