@@ -66,8 +66,8 @@ class Global extends LookUpBinding {
 
     s.write(
         "late final ${w.ffiLibraryPrefix}.Pointer<$cType> $pointerName = ${w.lookupFuncIdentifier}<$cType>('$originalName');\n\n");
-    if (type.broadType == BroadType.Struct) {
-      if (type.struc!.isOpaque) {
+    if (type.broadType == BroadType.Compound) {
+      if (type.compound!.isOpaque) {
         s.write(
             '${w.ffiLibraryPrefix}.Pointer<$cType> get $globalVarName => $pointerName;\n\n');
       } else {
