@@ -46,6 +46,7 @@ const includeCStdLib = 'include-c-standard-library';
 // Declarations.
 const functions = 'functions';
 const structs = 'structs';
+const unions = 'unions';
 const enums = 'enums';
 const unnamedEnums = 'unnamed-enums';
 const globals = 'globals';
@@ -58,10 +59,10 @@ const rename = 'rename';
 const memberRename = 'member-rename';
 const symbolAddress = 'symbol-address';
 
-const structDependencies = 'dependency-only';
-// Values for `structDependencies`.
-const fullStructDependencies = 'full';
-const opaqueStructDependencies = 'opaque';
+const dependencyOnly = 'dependency-only';
+// Values for `compoundDependencies`.
+const fullCompoundDependencies = 'full';
+const opaqueCompoundDependencies = 'opaque';
 
 const structPack = 'pack';
 const Map<Object, int?> packingValuesMap = {
